Change BitInt constructor from long to int
Underlying C/C++ libraries on Windows are likely to have 4 byte longs,
so use Java ints rather than longs.

+++ zillions-gmp/src/main/java/uk/co/magictractor/zillions/gmp/GmpBigInt.java
@@ -36,7 +36,7 @@ public class GmpBigInt implements BigInt {
         }
     }
 
-    public GmpBigInt(long x) {
+    public GmpBigInt(int x) {
         __lib.mpz_init_set_si(_mpz, new NativeLong(x));
     }
